// src/app/shared/models/gerencia.ts
export class Gerencia {

    idGerencia?:number;
    descripcionGerencia?:string;
    idSede?:number;
    idEmpresa?:number;
    indicadorBaja?: string;
    usuarioCreacion?: string;
    ipCreacion?: string;
    fechaCreacion?: string;
    usuarioModificacion?: string;
    ipModificacion?: string;
    fechaModificacion?: string;
}